fix certain breaking stuff

+++ plasm2_emu/cpu/ops/cops_interrupts.c
@@ -50,7 +50,7 @@ void IRT(void) {
 	union {
 		u64 Raw;
 		struct {
-			u64 Flags;
+			u32 Flags;
 			byte SecurityLevel;
 			byte CallFlag;
 			u16 Reserved;
